rename: 폴더 이름 카멜케이스로 변경
